Remove process_manager views tests from main views tests
